refactor: extract business logic from JSX components into custom hooks
Ran two passes of logic/UI separation across the client codebase.

Pass 1 — generic components:
- Signup.jsx → useSignupForm (useState + useEffect + async submit)
- TableWithSort.jsx → useTableWithSort (useState + useMemo sorting)
- ViewReviews.jsx — moved auto-dismiss useEffect into useReviewsHandlers
- Form.jsx → useFormState (useState + handleChange + handleFormSubmit)
- OpenModal.jsx → useOpenModalKeyboard (useEffect ESC key listener)

Pass 2 — table row construction (Category D):
- MessagesTable.jsx → useMessagesTableRows (trTd with isAdmin branch + date formatting)
- ServiceAdminTable.jsx → useServiceAdminTableRows (35-line trTd with conditional classes)
- MsgOfContactTable.jsx → useMsgOfContactTableRows (trTd with date formatting)
- CarsTable.jsx → useCarsTableRows
- AccountTable.jsx → useAccountTableRows
- UsersTable.jsx → useUsersTableRows (trTh with dynamic sort handlers)
- AccountServices.jsx → useAccountServicesData (extracted direct store reads)
- AppointmentCalendar.jsx → useAppointmentCalendar (useMemo event transformation)

~280 logic lines removed from components; 13 new hook files created.
