+++ client/src/components/landing/reviews/ViewReviews.jsx
@@ -1,4 +1,3 @@
-import { useEffect } from "react";
 import SwiperReviews from "./SwiperReviews";
 import AddReview from "./AddReview";
 import Submitted from "../../Submitted";
@@ -6,12 +5,6 @@ import { useReviewsState as useReviewsContext } from "./hooks/useReviewsState";
 
 export default function ViewReviews() {
   const { isSubmittedTemp, setIsSubmittedTemp } = useReviewsContext();
-
-  useEffect(() => {
-    if (!isSubmittedTemp) return;
-    const id = setTimeout(() => setIsSubmittedTemp(false), 7000);
-    return () => clearTimeout(id);
-  }, [isSubmittedTemp, setIsSubmittedTemp]);
 
   if (isSubmittedTemp) {
     return (
