+++ client/src/components/openModal/hooks/useOpenModalKeyboard.js
@@ -1,0 +1,12 @@
+import { useEffect } from "react";
+
+export function useOpenModalKeyboard(isOpen, onClose) {
+  useEffect(() => {
+    if (!isOpen || !onClose) return;
+    const handleKeyDown = (e) => {
+      if (e.key === "Escape") onClose();
+    };
+    document.addEventListener("keydown", handleKeyDown);
+    return () => document.removeEventListener("keydown", handleKeyDown);
+  }, [isOpen, onClose]);
+}
