+++ client/src/components/table/TableWithSort.jsx
@@ -1,50 +1,9 @@
 import "./table.css";
-import { Children, cloneElement, useState, useMemo } from "react";
 import PropTypes from "prop-types";
+import { useTableWithSort } from "./hooks/useTableWithSort";
 
 const Table = ({ trTh, trTd }) => {
-  const [sortConfig, setSortConfig] = useState({
-    key: null,
-    direction: "ascending",
-  });
-
-  const handleSort = (key) => {
-    setSortConfig((prevConfig) => ({
-      key,
-      direction:
-        prevConfig.key === key && prevConfig.direction === "ascending" ? "descending" : "ascending",
-    }));
-  };
-
-  const sortedData = useMemo(() => {
-    if (!sortConfig.key || !trTd) return trTd;
-
-    return [...trTd].sort((a, b) => {
-      const aValue = a.props.children.find((child) => child.props["data-label"] === sortConfig.key)
-        ?.props.children;
-      const bValue = b.props.children.find((child) => child.props["data-label"] === sortConfig.key)
-        ?.props.children;
-
-      const aDate = Date.parse(aValue);
-      const bDate = Date.parse(bValue);
-      const aNum = isNaN(aDate) ? aValue : aDate;
-      const bNum = isNaN(bDate) ? bValue : bDate;
-
-      if (aNum < bNum) return sortConfig.direction === "ascending" ? -1 : 1;
-      if (aNum > bNum) return sortConfig.direction === "ascending" ? 1 : -1;
-      return 0;
-    });
-  }, [sortConfig, trTd]);
-
-  const columnCount = trTh?.props?.children?.length || 1;
-
-  const headers = Children.map(trTh?.props?.children, (child) =>
-    cloneElement(child, {
-      onClick: () => handleSort(child.props.children),
-      style: { cursor: "pointer" },
-      className: sortConfig.key === child.props.children ? `sorted-${sortConfig.direction}` : "",
-    })
-  );
+  const { sortedData, headers, columnCount } = useTableWithSort(trTh, trTd);
 
   return (
     <section className="table__body">
