+++ client/src/pages/messagesOfContact/MsgOfContactTable.jsx
@@ -1,42 +1,10 @@
 import "../../components/table/table.css";
 import Search from "../../components/table/Search";
 import Table from "../../components/table/TableWithSort";
-import { getMomentFromUpdatedAt } from "../../utils/dateUtils";
-import { useMsgOfContactTable } from "./hooks/useMsgOfContactTable";
+import { useMsgOfContactTableRows } from "./hooks/useMsgOfContactTableRows";
 
 export default function MsgOfContactTable() {
-  const { displayContacts, handleSearch, handleContact, handleExport } = useMsgOfContactTable();
-
-  const trTh = (
-    <tr>
-      <th></th>
-      <th>First Name</th>
-      <th>Last Name</th>
-      <th>Email</th>
-      <th>Phone</th>
-      <th>Message</th>
-      <th>Date</th>
-    </tr>
-  );
-
-  const trTd = displayContacts?.map((message) => {
-    const { theDate } = getMomentFromUpdatedAt(message.updatedAt);
-    return (
-      <tr key={message._id}>
-        <td>
-          <button value={message._id} name="deleteContact" onClick={handleContact}>
-            Delete
-          </button>
-        </td>
-        <td data-label="First Name">{message.firstName}</td>
-        <td data-label="Last Name">{message.lastName}</td>
-        <td data-label="Email">{message.email}</td>
-        <td data-label="Phone">{message.phone}</td>
-        <td data-label="Message">{message.message}</td>
-        <td data-label="Date">{theDate}</td>
-      </tr>
-    );
-  });
+  const { trTh, trTd, handleSearch, handleExport } = useMsgOfContactTableRows();
 
   return (
     <div className="table-container">
