+++ client/src/pages/servicesAdmin/hooks/useServiceAdminTableRows.jsx
@@ -1,0 +1,54 @@
+import { useServiceAdminTableData } from "./useServiceAdminTableData";
+
+export function useServiceAdminTableRows() {
+  const { displayServices, handleSearch, handleServiceIdAction, handleExport } =
+    useServiceAdminTableData();
+
+  const trTh = (
+    <tr>
+      <th></th>
+      <th>car</th>
+      <th>title</th>
+      <th>description</th>
+      <th>price</th>
+      <th>paid</th>
+      <th>status</th>
+    </tr>
+  );
+
+  const trTd = displayServices?.map((service) => (
+    <tr key={service?._id}>
+      <td data-label="Actions">
+        <button name="manage" value={service?._id} onClick={handleServiceIdAction}>
+          Manage
+        </button>
+      </td>
+      <td data-label="Car">{service?.car?.numberPlate}</td>
+      <td data-label="Title">{service?.title}</td>
+      <td data-label="Description">{service?.description}</td>
+      <td data-label="Price">{service?.price}</td>
+      <td data-label="Payment Status">
+        <button
+          name="editPaid"
+          value={service?._id}
+          onClick={handleServiceIdAction}
+          className={`editPaid ${service?.paid ? "Paid" : "Unpaid"}`}
+        >
+          {service?.paid ? "Paid" : "Unpaid"}
+        </button>
+      </td>
+      <td data-label="Status">
+        <button
+          className={`status ${service?.status}`}
+          name="editStatus"
+          value={service?._id}
+          onClick={handleServiceIdAction}
+        >
+          {service?.status}
+        </button>
+      </td>
+    </tr>
+  ));
+
+  return { trTh, trTd, handleSearch, handleExport };
+}
